chore: move SetCwdToFileDir to new package, workdir, so it can be used in the file domain
