+++ src/pkg/common/common_test.go
@@ -3,7 +3,6 @@ package common_test
 import (
 	"context"
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
@@ -280,65 +279,6 @@ func TestValidationFromString(t *testing.T) {
 		})
 	}
 
-}
-
-func TestSwitchCwd(t *testing.T) {
-
-	tempDir := t.TempDir()
-
-	tests := []struct {
-		name     string
-		path     string
-		expected string
-		wantErr  bool
-	}{
-		{
-			name:     "Valid path",
-			path:     tempDir,
-			expected: tempDir,
-			wantErr:  false,
-		},
-		{
-			name:     "Path is File",
-			path:     "./common_test.go",
-			expected: "./",
-			wantErr:  false,
-		},
-		{
-			name:     "Invalid path",
-			path:     "/nonexistent",
-			expected: "",
-			wantErr:  true,
-		},
-		{
-			name:     "Empty Path",
-			path:     "",
-			expected: "",
-			wantErr:  true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			resetFunc, err := common.SetCwdToFileDir(tt.path)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("SwitchCwd() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if err == nil {
-				defer resetFunc()
-				wd, _ := os.Getwd()
-				expected, err := filepath.Abs(tt.expected)
-				if err != nil {
-					t.Errorf("SwitchCwd() error = %v, wantErr %v", err, tt.wantErr)
-					return
-				}
-				if !strings.HasSuffix(wd, expected) {
-					t.Errorf("SwitchCwd() working directory = %v, want %v", wd, tt.expected)
-				}
-			}
-		})
-	}
 }
 
 func TestValidationToResource(t *testing.T) {
